+++ profile_writer_test.go
@@ -70,6 +70,36 @@ func TestProfileWriter_WriteProfile(t *testing.T) {
 	assertKeyValue(t, "aws_session_token", "some-session-token", content)
 }
 
+func TestProfileWriter_WriteProfileWoRegion(t *testing.T) {
+	credPath := path.Join(os.TempDir(), "swamp-test.ini")
+	os.Remove(credPath)
+
+	os.Setenv("AWS_SHARED_CREDENTIALS_FILE", credPath)
+	defer os.Clearenv()
+	defer os.Remove(credPath)
+
+	profileName := "some-profile"
+	region := ""
+	creds := &sts.Credentials{}
+	creds.SetAccessKeyId("some-access-key")
+	creds.SetSecretAccessKey("some-secret-access-key")
+	creds.SetSessionToken("some-session-token")
+
+	pw, _ := NewProfileWriter()
+	pw.WriteProfile(creds, &profileName, &region)
+
+	b, err := ioutil.ReadFile(credPath)
+	assert.NoError(t, err)
+
+	content := string(b)
+
+	assert.Regexp(t, `^\[some-profile\]\n.*`, content)
+	assert.NotRegexp(t, `region`, content)
+	assertKeyValue(t, "aws_access_key_id", "some-access-key", content)
+	assertKeyValue(t, "aws_secret_access_key", "some-secret-access-key", content)
+	assertKeyValue(t, "aws_session_token", "some-session-token", content)
+}
+
 func assertKeyValue(t *testing.T, key, value, content string) {
 	assert.Regexp(t, fmt.Sprintf(`\n%s\s*=\s*%s\n.*`, key, value), content)
 }